Fixed bug in Selenium2Driver::reset() method.

+++ src/Selenium2Driver.php
@@ -306,6 +306,10 @@ class Selenium2Driver extends CoreDriver
      */
     public function reset()
     {
+        if (!$this->isStarted()) {
+            return;
+        }
+
         $this->webDriver->manage()->deleteAllCookies();
     }
